More refactoring Sketchup vs SVG output
